refactored queries storage
